#7 - some refactoring
