--- triss-core/src/main/kotlin/com/pp/trisscore/model/FinancialSource.kt
@@ -1,5 +1,0 @@
-package com.pp.trisscore.model
-
-class FinancialSource {
-
-}
