--- triss-core/src/main/kotlin/com/pp/trisscore/model/IdentityDocument.kt
@@ -1,13 +1,0 @@
-package com.pp.trisscore.model
-
-/**
- *
- * author: Marceli Jerzyński
- * e-mail: marceli.jerzynski@student.put.poznan.pl
- * Date: 03.10.2020
- **/
-data class IdentityDocument (
-    val Type: DocumentType,
-    val Name: String,
-    val Number: String
-)
