--- triss-core/src/main/kotlin/com/pp/trisscore/model/Prepayment.kt
@@ -1,12 +1,0 @@
-package com.pp.trisscore.model
-
-/**
- *
- * author: Marceli Jerzyński
- * e-mail: marceli.jerzynski@student.put.poznan.pl
- * Date: 03.10.2020
- **/
-data class Prepayment (
-    val conferenceFee: Fee?,
-    val accommodation: Fee?
-)
