+++ triss-core/src/main/kotlin/com/pp/trisscore/model/classes/AdvanceRequest.kt
@@ -1,4 +1,4 @@
-package com.pp.trisscore.model
+package com.pp.trisscore.model.classes
 
 import java.sql.Date
