+++ triss-core/src/main/kotlin/com/pp/trisscore/model/classes/Employee.kt
@@ -1,7 +1,12 @@
-package com.pp.trisscore.model
+package com.pp.trisscore.model.classes
 
 import java.sql.Date
-
+/**
+ *
+ * author: Marceli Jerzyński
+ * e-mail: marceli.jerzynski@student.put.poznan.pl
+ * Date: 03.10.2020
+ **/
 data class Employee (
         val id: Int,
         val firstName: String,
