+++ triss-core/src/main/kotlin/com/pp/trisscore/model/classes/Fee.kt
@@ -1,4 +1,6 @@
-package com.pp.trisscore.model
+package com.pp.trisscore.model.classes
+
+import com.pp.trisscore.model.enums.PaymentType
 
 /**
  *
