+++ triss-core/src/main/kotlin/com/pp/trisscore/model/classes/FinancialSource.kt
@@ -1,0 +1,5 @@
+package com.pp.trisscore.model.classes
+
+class FinancialSource { //TODO
+
+}
