+++ triss-core/src/main/kotlin/com/pp/trisscore/model/classes/Place.kt
@@ -1,6 +1,4 @@
-package com.pp.trisscore.model
-
-import java.util.*
+package com.pp.trisscore.model.classes
 
 /**
  *
